requirements.txt
feedparser
requests
flask
pandas
python-dotenv
google-genai
